Bumped loupe submodule
- Reflected enums are now based on uint16_t types

+++ gemcutter/Rendering/Light.h
@@ -10,7 +10,7 @@ namespace gem
 	class Light : public Component<Light>
 	{
 	public:
-		enum class Type : unsigned
+		enum class Type : uint16_t
 		{
 			Point,
 			Directional,
